Packaging of data/graph-classes.
