+++ stargraph-core/src/main/resources/reference.conf
@@ -1673,7 +1673,7 @@ stargraph = {
     triple-store.hdt.use-index = no
 
     model.facts = {
-      provider.class = net.stargraph.core.FactProviderFactory
+      provider.class = net.stargraph.core.data.FactProviderFactory
 
       processors = [
         {
@@ -1691,7 +1691,7 @@ stargraph = {
     }
 
     model.entities = {
-      provider.class = net.stargraph.core.EntityProviderFactory
+      provider.class = net.stargraph.core.data.EntityProviderFactory
       index-store.factory.class = net.stargraph.core.impl.lucene.LuceneFactory
 
       processors = [
@@ -1706,7 +1706,7 @@ stargraph = {
     }
 
     model.relations = {
-      provider.class = net.stargraph.core.PropertyProviderFactory
+      provider.class = net.stargraph.core.data.PropertyProviderFactory
 
       processors = [
         {
@@ -1727,7 +1727,7 @@ stargraph = {
     }
 
     model.documents = {
-      provider.class = net.stargraph.core.DocumentProviderFactory
+      provider.class = net.stargraph.core.data.DocumentProviderFactory
 
       processors = [
         {
@@ -1757,7 +1757,7 @@ stargraph = {
     triple-store.hdt.use-index = no
 
     model.facts = {
-      provider.class = net.stargraph.core.FactProviderFactory
+      provider.class = net.stargraph.core.data.FactProviderFactory
 
       processors = [
         {
@@ -1775,7 +1775,7 @@ stargraph = {
     }
 
     model.entities = {
-      provider.class = net.stargraph.core.EntityProviderFactory
+      provider.class = net.stargraph.core.data.EntityProviderFactory
       # We chose to use Lucene library directly. If this is ommitted the global configuration is used.
       index-store.factory.class = net.stargraph.core.impl.lucene.LuceneFactory
 
@@ -1791,7 +1791,7 @@ stargraph = {
     }
 
     model.relations = {
-      provider.class = net.stargraph.core.PropertyProviderFactory
+      provider.class = net.stargraph.core.data.PropertyProviderFactory
 
       processors = [
         {
@@ -1812,7 +1812,7 @@ stargraph = {
     }
 
     model.documents = {
-      provider.class = net.stargraph.core.DocumentProviderFactory
+      provider.class = net.stargraph.core.data.DocumentProviderFactory
 
       processors = [
         {
