+++ stargraph-core/src/test/java/net/stargraph/test/SimpleKBTest.java
@@ -28,7 +28,7 @@ package net.stargraph.test;
 
 import com.typesafe.config.Config;
 import com.typesafe.config.ConfigFactory;
-import net.stargraph.core.NTriplesModelFactory;
+import net.stargraph.core.impl.ntriples.NTriplesModelFactory;
 import net.stargraph.core.Stargraph;
 import net.stargraph.core.index.Indexer;
 import net.stargraph.core.index.NullIndicesFactory;
