+++ stargraph-core/src/test/resources/application.conf
@@ -43,7 +43,7 @@ stargraph = {
     }
 
     model.facts = {
-      provider.class = net.stargraph.core.FactProviderFactory
+      provider.class = net.stargraph.core.data.FactProviderFactory
 
       processors = [
         {
@@ -64,7 +64,7 @@ stargraph = {
     }
 
     model.entities = {
-      provider.class = net.stargraph.core.EntityProviderFactory
+      provider.class = net.stargraph.core.data.EntityProviderFactory
       processors = [
         {
           sink-duplicate: ${processor.sink-duplicate}
